linux: Add desktop-entry and AppStream metainfo files
The desktop-entry file is needed to give the application an icon on
Linux and also make it launchable from the menu.
The metainfo file describes general metadata about the application and
makes it show up in software centers and catalogues. It also registers a
unique name ("com.intantech.intan_rhx") for the software, so it can be
identified by the system.
Validate this file via:
  appstreamcli validate com.intantech.intan_rhx.metainfo.xml