linux: Add basic Debian/Ubuntu packaging recipe
This is the bare minimum needed for building a Debian/Ubuntu package
from source.
Some hacks via source/options were required in order to make the build
process not delete the prebuilt OpalKelly shared library.
The package is unfit for Debian itself, but for casual automated builds
and separate distribution it's good enough.

Build by copyring the debian/ directory to the root source tree, then
run `debuild` to get a package.